requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
jinja2
python-multipart
